Move world ranking to own class
