+++ src/Domain/Ranking/IFSCWorldRankingException.php
@@ -1,0 +1,14 @@
+<?php declare(strict_types=1);
+
+/**
+ * @license  http://opensource.org/licenses/mit-license.php MIT
+ * @link     https://github.com/nicoSWD
+ * @author   Nicolas Oelgart <nico@oelgart.com>
+ */
+namespace nicoSWD\IfscCalendar\Domain\Ranking;
+
+use Exception;
+
+final class IFSCWorldRankingException extends Exception
+{
+}
